requirements.txt
asgiref==3.8.1

attrs==25.3.0

awsebcli==3.21.0

beautifulsoup4==4.13.3

#boto3==1.37.24

#botocore==1.37.24

botocore>=1.35.0,<1.36.0

cachetools==5.5.2

cassandra-driver==3.29.2

cement>=2.10.14,<3.0.0
#cement==3.0.12

certifi==2025.1.31

channels==4.2.2

channels-redis==4.2.1

charset-normalizer==3.2.0

click==8.1.7

colorama==0.4.6

constantly==23.10.4

coreapi==2.3.3

coreschema==0.0.4

crispy-bootstrap4==2024.10

crispy-bootstrap5==2024.10

cryptography==44.0.2

daphne==4.1.2

decorator==5.2.1

defusedxml==0.7.1

diff-match-patch==20241021

dj-database-url==2.3.0

Django==4.2.0

django-admin-sortable2==2.2.5

django-anymail==12.0

django-bootstrap4==25.1

django-ckeditor==6.7.2

django-ckeditor-5==0.2.10

django-cors-headers==4.7.0

django-crispy-forms==2.3

django-dbbackup==4.2.1

django-dotenv==1.4.2

django-environ==0.12.0

django-import-export==4.3.7

django-jazzmin==3.0.1

django-recaptcha==4.1.0

django-rest-auth==0.9.5

djangorestframework==3.16.0

django-six==1.0.5

django-static-fontawesome==6.5.2.0

django-storages==1.14.5

django-user-agents==0.4.0

djangorestframework-simplejwt==5.5.0

drf-yasg==1.21.10

environs==14.1.1

geoip2==5.0.1

geomet>=0.1,<0.3

google-api-core==2.24.2

google-api-python-client==2.166.0

google-auth==2.38.0

google-auth-httplib2==0.2.0

google-auth-oauthlib==1.2.1

googleapis-common-protos==1.69.2

gunicorn==23.0.0

httplib2==0.22.0

humanize==4.12.2

hyperlink==21.0.0

idna==3.10

imageio==2.37.0

imageio-ffmpeg==0.6.0

Jinja2==3.1.6

Markdown==3.7

MarkupPy==1.18

MarkupSafe==3.0.2

marshmallow==3.26.1

maxminddb==2.6.3

moviepy==2.1.2

msal==1.32.0

msgpack==1.1.0

numpy==2.2.4

oauthlib==3.2.2

packaging==24.2

pandas==2.2.3

prolog==0.3.1

proto-plus==1.26.1

protobuf==6.30.2

psycopg2-binary==2.9.10

pyasn1==0.6.1

pyasn1-modules==0.4.2

pycountry==24.6.1

PyJWT>=1.7.1,<2.10.0

python-dotenv==1.1.0

python3-openid==3.2.0

pytz==2025.2

PyYAML==6.0.2

redis==5.2.1

requests==2.32.3

requests-oauthlib==2.0.0

rjsmin==1.2.4

rsa==4.9

s3transfer>=0.6.0 #updated

service-identity==24.2.0

setuptools>=65.0.0

shortuuid==1.0.13

six==1.17.0

sqlparse==0.5.3

stripe==11.6.0

tqdm==4.67.1

twilio==9.5.1

tzdata==2025.2

uritemplate==4.1.1

urllib3>=1.26.5,<2.0

whitenoise==6.9.0

django-extensions==3.2.3

razorpay==1.4.2